--- Views/Atualizar/Atualizar.cshtml
@@ -1,20 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Atualizar";
-}
-<h1>@ViewData["Title"]</h1>
-
-<form method="post">
-    <label>CPF :</label>
-    <input type="text" id="id" name="cpf" />
-    <label>Nome :</label>
-    <input type="text" id="id" name="nome" />
-    <br />
-    <label>telefone:</label>
-    <input type="text" id="id" name="telefone" />
-    <br />
-    <label>Data de Nascimento:</label>
-    <input type="text" id="id" name="dataNascimento" />
-    <br />
-    <button>Alterar</button>
-</form>
-
