+++ Views/Cliente/Create.cshtml
@@ -1,11 +1,11 @@
 ﻿@{
-    ViewData["Title"] = "Cadastro";
+    ViewData["Title"] = "Create";
 }
 <h1>Cadastrar Clientes</h1>
 
 <div class="row">
     <div class="col-md-4">
-        <form asp-action="Cadastro" method="post">
+        <form asp-action="Create" method="post">
             <div asp-validation-summary="All" class="text-danger"></div>
             <div class="form-group">
                 <label class="control-label">CPF</label>
@@ -33,5 +33,4 @@
         </form>
     </div>
 </div>
-
