+++ Views/Cliente/Delete.cshtml
@@ -1,0 +1,17 @@
+﻿@{
+    ViewData["Title"] = "Delete";
+}
+
+<h1>@ViewData["Title"]</h1>
+
+<form method="post">
+    <div class="form-group">
+        <label class="control-label">CPF</label>
+        <input type="text" class="form-control" id="CPF" name="cpfCliente" required>
+        <span class="text-danger"></span>
+    </div>
+    <div class="form-group">
+        <input type="submit" value="Delete" class="btn btn-primary" />
+    </div>
+</form>
+
