+++ Views/Cliente/Read.cshtml
@@ -1,7 +1,7 @@
 ﻿@model List<PizzaEverisDay.Models.Cliente>
 
 @{
-    ViewData["Title"] = "Clientes";
+    ViewData["Title"] = "Read";
 }
 
 <table class = "table">
