+++ Views/Cliente/Update.cshtml
@@ -1,0 +1,36 @@
+﻿@{
+    ViewData["Title"] = "Update";
+}
+<h1>@ViewData["Title"]</h1>
+
+
+<div class="row">
+    <div class="col-md-4">
+        <form asp-action="Update" method="post">
+            <div asp-validation-summary="All" class="text-danger"></div>
+            <div class="form-group">
+                <label class="control-label">CPF</label>
+                <input type="text" class="form-control" id="CPF" name="cpf" required>
+                <span class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label class="control-label">Nome Completo</label>
+                <input type="text" class="form-control" id="Nome" name="nome" required>
+                <span asclass="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label class="control-label">Dada de nascimento</label>
+                <input type="text" class="form-control" id="Data_Nascimento" name="dataNascimento" required>
+                <span class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <label class="control-label">Telefone</label>
+                <input type="text" class="form-control" id="Telefone" name="telefone" required>
+                <span class="text-danger"></span>
+            </div>
+            <div class="form-group">
+                <input type="submit" value="Atualizar" class="btn btn-primary" />
+            </div>
+        </form>
+    </div>
+</div>
