--- Views/Deletar/Deletar.cshtml
@@ -1,11 +1,0 @@
-﻿    @{
-    ViewData["Title"] = "Deletar";
-}
-
-<h1>@ViewData["Title"]</h1>
-
-<form method="post">
-    <label>CPF do Cliente para excluir:</label>
-    <input type="text" id="id" name="cpfCliente" />
-    <button>Deletar</button>
-</form>
