+++ Views/Error/ErroCliente.cshtml
@@ -1,0 +1,7 @@
+﻿<html>
+<head>
+    <script>
+        alert("CPF não encontrado no banco de dados.")
+    </script>
+</head>
+</html>
