+++ Views/Shared/_Layout.cshtml
@@ -23,19 +23,9 @@
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
-                        </li>
+                        </li> 
                         <li>
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Cadastrar" asp-action="Cadastro">Cadastrar</a>
-                        </li>
-                        <li>
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Atualizar" asp-action="Atualizar">Atualizar Cadastro</a>
-                        </li>
-                        <li>
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Deletar" asp-action="Deletar">Deletar Cadastro</a>
-
-                        </li>
-                        <li>
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Listar" asp-action="Listar">Lista De Clientes</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Cliente" asp-action="Read">Lista De Clientes</a>
                         </li>
 
                     </ul>
